pyproject.toml
[tool.poetry]
name = "basic-django"
version = "0.1.0"
description = ""
authors = ["Ibrahim Khalid <ibrahim.khalid@touristsexpi.tz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
djangorestframework = "^3.14.0"
pygments = "^2.16.1"
pillow = "^10.0.1"
drf-spectacular = "^0.26.5"
djangorestframework-simplejwt = "^5.3.0"
gunicorn = "^21.2.0"
django-cors-headers = "^4.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
